Implement Object.ToString, define System.Int64

+++ stdlib/corlib/Int64.cs
@@ -1,16 +1,16 @@
 namespace System
 {
     /// <summary>
-    /// Represents a 32-bit integer.
+    /// Represents a 64-bit integer.
     /// </summary>
-    public struct Int32 : Object, IEquatable<Int32>
+    public struct Int64 : Object, IEquatable<Int64>
     {
         // Note: integers are equivalent to instances of this data structure because
         // flame-llvm stores the contents of single-field structs as a value of their
-        // field, rather than as an LLVM struct. So a 32-bit integer becomes an i32 and
-        // so does a `System.Int32`. So don't add, remove, or edit the fields in this
+        // field, rather than as an LLVM struct. So a 64-bit integer becomes an i64 and
+        // so does a `System.Int64`. So don't add, remove, or edit the fields in this
         // struct.
-        private int value;
+        private long value;
 
         /// <summary>
         /// Converts this integer to a string representation.
@@ -22,7 +22,7 @@ namespace System
         }
 
         /// <inheritdoc/>
-        public bool Equals(Int32 other)
+        public bool Equals(Int64 other)
         {
             return value == other.value;
         }
@@ -30,13 +30,13 @@ namespace System
         /// <inheritdoc/>
         public sealed override bool Equals(Object other)
         {
-            return other is Int32 && Equals((Int32)other);
+            return other is Int64 && Equals((Int64)other);
         }
 
         /// <inheritdoc/>
         public sealed override int GetHashCode()
         {
-            return value;
+            return (int)value;
         }
     }
 }
